requirement.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
backports.zoneinfo==0.2.1
certifi==2022.9.14
cffi==1.15.1
charset-normalizer==2.1.1
coverage==6.4.4
cryptography==38.0.1
Django==4.1.1
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
h11==0.13.0
idna==3.4
mysqlclient==2.1.1
numpy==1.23.3
outcome==1.2.0
pandas==1.4.4
pycparser==2.21
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
selenium==4.4.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tqdm==4.64.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.3
wsproto==1.2.0
